Zmiana przekazywania informacj na dto, poprawa nazewnictwa

+++ backend/src/main/java/com/adamlewandowski/repository/WeatherInformationRepository.java
@@ -17,7 +17,7 @@ public interface WeatherInformationRepository extends JpaRepository<WeatherDao, 
     List<WeatherDao> searchForOneCityOnePage(@Param("searchTerm") String searchTerm, @Param("startRow") int startRow, @Param("numberOfRowsToDisplay") int numberOfRowsToDisplay);
 
     @Query(value = "SELECT count(*) FROM weather WHERE city_name LIKE lower(concat('%', :searchTerm,'%'))", nativeQuery = true)
-    long checkNumberOfElements(@Param("searchTerm") String searchTerm);
+    int checkNumberOfElements(@Param("searchTerm") String searchTerm);
 
     @Query(value = "SELECT * FROM weather limit  :startRow, :numberOfRowsToDisplay", nativeQuery = true)
     List<WeatherDao> searchForAllCicites(@Param("startRow") int startRow, @Param("numberOfRowsToDisplay") int numberOfRowsToDisplay);
